add stricter compilation for auto_mapping_ros src

+++ auto_mapping_ros/node/coverage_sequence_csv_creator.cpp
@@ -35,7 +35,7 @@ int main() {
   // Find Best Sequence
   const auto sequences_node = aco::solve_vrp(aco_graph, init_id);
 
-  for (int i = 0; i < sequences_node.first.size(); i++) {
+  for (size_t i = 0; i < sequences_node.first.size(); i++) {
     std::vector<std::array<int, 2>> current_sequence{};
     for (const auto& node : sequences_node.first[i]) {
       current_sequence.emplace_back(std::array<int, 2>{
